Creates sub folders for every node group
