+++ nodes/generators/generator_point.py
@@ -38,7 +38,7 @@ class Point(FCNNode):
 
     icon: str = icon("fcn_default.png")
     op_title: str = "Point"
-    op_category = "Generator"
+    op_category: str = "Generator"
     content_label_objname: str = "fcn_node_bg"
 
     def __init__(self, scene):
