+++ nodes/generators/generator_solid_box.py
