+++ nodes/number/number_number.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 ###################################################################################
 #
-#  scene_get_objects_data.py
+#  number_number.py
 #
 #  Copyright (c) 2022 Ronny Scharf-Wildenhain <ronny.scharf08@gmail.com>
 #
@@ -22,49 +22,40 @@
 #
 #
 ###################################################################################
-import FreeCAD
+from decimal import Decimal
+
+from qtpy.QtWidgets import QLineEdit
+from nodeeditor.node_content_widget import QDMNodeContentWidget
 
 from fcn_conf import register_node
-from fcn_base_node import FCNNode
+from fcn_base_node import FCNNode, FCNNodeContentView
 from fcn_locator import icon
 
 
 @register_node
-class GetObjectsData(FCNNode):
+class Number(FCNNode):
 
     icon: str = icon("fcn_default.png")
-    op_title: str = "Get Objects Data"
-    op_category = "Scene"
+    op_title: str = "Number"
+    op_category: str = "Number"
     content_label_objname: str = "fcn_node_bg"
 
     def __init__(self, scene):
         super().__init__(scene=scene,
-                         inputs_init_list=[(3, "In", 1, "Object label", True, ("str", ))],
-                         outputs_init_list=[(4, "Obj", 0, 0, True, ("fc_obj", ))],
-                         width=170)
+                         inputs_init_list=[(0, "In", 1, 0, False, ('int', 'float'))],
+                         outputs_init_list=[(0, "Out", 0, 0, True, ('int', 'float'))],
+                         width=150)
 
     def collapse_node(self, collapse: bool = False):
         super().collapse_node(collapse)
 
-        if collapse is True:
-            self.title = 'Obj: ' + str(self.sockets_input_data[0][0])
+        if (collapse is True) and isinstance(self.sockets_input_data[0][0], (int, float)):
+            self.title: str = 'In: %.2E' % Decimal(str(self.sockets_input_data[0][0]))
         else:
-            self.title = self.default_title
+            self.title: str = self.default_title
 
     def eval_operation(self, sockets_input_data: list) -> list:
         self.collapse_node(self.content.isHidden())
 
-        label_list: list = sockets_input_data[0]
-        obj_list: list = []
-
-        if not (FreeCAD.ActiveDocument is None):
-            for label in label_list:
-                obj = FreeCAD.ActiveDocument.getObjectsByLabel(label)
-                if len(obj) == 1:
-                    obj_list.append(obj[0])
-                else:
-                    raise ValueError('Unknown object')
-        else:
-            raise Exception('No active document')
-
-        return [obj_list]
+        in_val: float = float(sockets_input_data[0][0])
+        return [[in_val]]
